Add files via upload
This is a piece of C# code from a game project I am working on with a friend. We needed a box collider trigger to check if the player had the correct key (hasKey == true), then destroy the trigger and open the two doors, allowing the player to leave the room. All code came from Unity documentation, just trial and error until we could get it to work.