Add files via upload
This script is from one of my game projects. It is the beginning of AI behavior for a security guard. The cop will pick a random patrol point, travel to it, and then pick the next point. This script was adapted from multiple 2D AI behavior tutorials, but they were intended for 2D platformers similar to Mario, which isn't applicable to my game. Through my own research in Unity's documentation, I was able to rotate the character image toward whatever the next selected point was.